google_route/requirements.txt
pycryptodomex>=3.19.0
ecdsa>=0.18.0
bleak>=0.21.0
aiohttp>=3.9.0
cryptography>=41.0.0
protobuf>=4.25.0
